requirements.txt
attrs==20.3.0
pyserial==3.5
wxPython==4.1.1
